refactor intrinsic module
